Visual C# console app to select a random name.
The names are characters from TV shows. Each name will be removed after it selected from the list.
Line 55- 60 are commented out will display the count of names already selected and list length as described in the comments above them if you uncomment these lines.
Feedback or ways to improve are always appreciated.